backend/requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
python-multipart==0.0.5
langchain==0.0.267
chromadb==0.3.29
pydantic>=1.9.0,<2.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==0.27.8
numpy==1.24.3
tokenizers==0.15.0